Makefile: add parallel make support

+++ repack-squashfs-mi.sh
@@ -11,21 +11,21 @@ set -e
 IMG=$1
 ROOTPW='$1$qtLLI4cm$c0v3yxzYPI46s28rbAYG//'  # "password"
 
-[ -e "$IMG" ] || { echo "rootfs img not found $IMG"; exit 1; }
+[ -e "$IMG" ] || { echo "rootfs img not found $IMG"; }
 
 # verify programs exist
 command -v unsquashfs &>/dev/null || { echo "install unsquashfs"; exit 1; }
 mksquashfs -version >/dev/null || { echo "install mksquashfs"; exit 1; }
 
-FSDIR=`mktemp -d /tmp/resquash-rootfs.XXXXX`
-trap "rm -rf $FSDIR" EXIT
+[ -z "$FSDIR" ] && FSDIR=`mktemp -d /tmp/resquash-rootfs.XXXXX`
+[ -z "$FSDIR" ] && trap "rm -rf $FSDIR" EXIT
 
 # test mknod privileges
-mknod "$FSDIR/foo" c 0 0 2>/dev/null || { echo "need to be run with fakeroot"; exit 1; }
+mknod "$FSDIR/foo" c 0 0 || { echo "need to be run with fakeroot"; exit 1; }
 rm -f "$FSDIR/foo"
 
 >&2 echo "unpacking squashfs..."
-unsquashfs -f -d "$FSDIR" "$IMG"
+[ -z "$FSDIR" ] && unsquashfs -f -d "$FSDIR" "$IMG"
 
 >&2 echo "patching squashfs..."
 
@@ -100,6 +100,9 @@ chown root:root "$FSDIR/sbin/xqflash"
 
 # as a last-ditch effort, change the *.miwifi.com hostnames to localhost
 #sed -i 's@\w\+.miwifi.com@localhost@g' $FSDIR/etc/config/miwifi
+
+# apply patch from xqrepack repository
+find patches -type f -exec bash -c "(cd "$FSDIR" && git apply -p1) < {}" \;
 
 >&2 echo "repacking squashfs..."
 rm -f "$IMG.new"
